backend/requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-webpack-loader==0.5.0
djangorestframework==3.9.1
Pillow==5.4.1
pytz==2018.9
whitenoise==4.1.2
